+++ db.json
@@ -2,11 +2,10 @@
   "users": [
     {
       "nome": "Luiz",
-      "Idade": "34",
-      "Gênero Favorito": "Ficção Científica",
+      "idade": 34,
+      "genero_favorito": "Ficção Científica",
       "estado": "PE",
-      "cidade":"Timbaúba"
-      
+      "cidade": "Timbaúba"
     }
   ],
   "livros": [
@@ -14,18 +13,27 @@
       "titulo": "O mundo de Sofia",
       "autor": "Jostein Gaarder",
       "genero": "Romance",
-      "sub-genero": "Filosofia",
+      "sub_genero": "Filosofia",
       "edicao": 1,
       "ano": 1991,
-      "editora":"AvaLivros"
+      "editora": "AvaLivros"
     },
     {
       "titulo": "Cem anos de solidão",
       "autor": "Gabriel García Marquez",
       "genero": "Romance",
-      "sub-genero": "Realismo Fantástico",
+      "sub_genero": "Realismo Fantástico",
       "edicao": 1,
       "ano": 1967,
+      "editora": "AvaLivros"
+    },
+    {
+      "titulo": "Pai Rico, Pai Pobre",
+      "autor": "Robert Kiyosaki",
+      "genero": "Finança pessoais",
+      "sub_genero": "Não ficção",
+      "edicao": 4,
+      "ano": 1997,
       "editora": "AvaLivros"
     }
   ]
